+++ apps/desktop/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.0.2](https://github.com/SubtractManufacturing/certtrace/compare/desktop-v1.0.1...desktop-v1.0.2) (2026-06-02)
+
+
+### Bug Fixes
+
+* Bump for release testing ([9e95054](https://github.com/SubtractManufacturing/certtrace/commit/9e95054ecedafdf01a46bede6c8095a8245b36a7))
+
+## [1.0.1](https://github.com/SubtractManufacturing/certtrace/compare/desktop-v1.0.0...desktop-v1.0.1) (2026-06-02)
+
+
+### Bug Fixes
+
+* Windows first-run errors and Create Library dark mode ([#24](https://github.com/SubtractManufacturing/certtrace/issues/24)) ([909f910](https://github.com/SubtractManufacturing/certtrace/commit/909f910838b142d7bdd40dde11ba8e4cbb2f38b3))
+
 ## 1.0.0 (2026-06-02)